Rename: user.ts로 파일명 변경
